Updates Vite themes setup documentation
Refactors the documentation to reflect the removal of the frontend theme configuration.

Simplifies the Vite configuration by reducing the number of separate configurations from three to two. Removes references to the future frontend theme, streamlines the development process and simplifies the documentation.
Also updates the ports to use the default Vite ports.